feat(config): modularise config
License: MIT
